ref DepartmentJDBCDao.class     DepartmentJDBCDao.class     DepartmentJDBCDao.java     DepartmentService.class     DepartmentService.class     DepartmentService.java     ICRUDService.class     ICRUDService.class     ICRUDService.java     IDepartmentDao.class     IDepartmentDao.class     IDepartmentDao.java     IDepartmentService.java

+++ src/main/java/by/it_academy/jd2/Mk_JD2_98_23/dao/api/ICRUDService.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface ICRUDService<T, S> {
     List<T> get();
 
-    T get(int id);
+    T get(long id);
 }
